graph,node,chain-ethereum: Move ETH env vars to chain crate
