+++ chain/ethereum/src/lib.rs
@@ -2,6 +2,7 @@ mod adapter;
 mod capabilities;
 pub mod codec;
 mod data_source;
+mod env;
 mod ethereum_adapter;
 mod ingestor;
 pub mod runtime;
@@ -11,6 +12,7 @@ pub use self::capabilities::NodeCapabilities;
 pub use self::ethereum_adapter::EthereumAdapter;
 pub use self::runtime::RuntimeAdapter;
 pub use self::transport::Transport;
+pub use env::ENV_VARS;
 
 // ETHDEP: These concrete types should probably not be exposed.
 pub use data_source::{DataSource, DataSourceTemplate, Mapping, MappingABI, TemplateSource};
