+++ graph/src/lib.rs
@@ -58,6 +58,7 @@ pub mod prelude {
     pub use async_trait::async_trait;
     pub use bigdecimal;
     pub use chrono;
+    pub use envconfig;
     pub use ethabi;
     pub use futures::future;
     pub use futures::prelude::*;
